+++ bindings/nodejs/lib/types/wallet/event.ts
@@ -85,12 +85,12 @@ class LedgerAddressGenerationWalletEvent extends WalletEvent {
 class NewOutputWalletEvent extends WalletEvent {
     private output: OutputData;
     private transaction?: TransactionPayload;
-    private transactionInputs?: [OutputResponse];
+    private transactionInputs?: OutputResponse[];
 
     constructor(
         output: OutputData,
         transaction?: TransactionPayload,
-        transactionInputs?: [OutputResponse],
+        transactionInputs?: OutputResponse[],
     ) {
         super(WalletEventType.NewOutput);
         this.output = output;
@@ -115,7 +115,7 @@ class NewOutputWalletEvent extends WalletEvent {
     /**
      * The transaction inputs.
      */
-    getTransactionInputs(): [OutputResponse] | undefined {
+    getTransactionInputs(): OutputResponse[] | undefined {
         return this.transactionInputs;
     }
 }